Move phase_spaces.jl into directory and split into files
